Adding keyboard shortcuts + fixing user repeated timeout calls
(previously fixed issue: keypress event in clockcontrols that were not detected)

Added shortcuts: space=start/stop, T=timeout, z=ctrl-z

Found and fixed issue: Timeout could be called multiple times in a row ressetting itself each time, it is unclear when the original clock restarted, so now the program can only call a timeout if it is not in timeout mode. If one tries to call a timeout during a timeout it ignores the call and prints a warning.